test: move timeout redaction cases into errors.test.ts
Keep error tests next to the other error-class coverage and exercise the constructor directly instead of a hanging fetch.
